+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to the Dullahan project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2024-10-01
+
+### Added
+
+- Command-line options for setup.sh script could be found by running with --help
+- Improved documentation for setup.sh script usage in README.md
+
+### Changed
+
+- Enhanced setup.sh script to pass additional arguments to the tracker's setup script
+- Updated README.md with new setup script options and usage examples
+
 ## [0.4.0] - 2024-09-29
 
 ### Added